+++ src/firewallfabrik/platforms/nftables/_nat_compiler.py
@@ -32,6 +32,7 @@ from firewallfabrik.compiler.processors._generic import (
     DropIPv6Rules,
     EmptyGroupsInRE,
     ExpandGroups,
+    RecursiveGroupsInRE,
     ResolveMultiAddress,
     SimplePrintProgress,
 )
@@ -111,6 +112,9 @@ class NATCompiler_nft(NATCompiler):
 
         self.add(ResolveMultiAddress('resolve compile-time MultiAddress'))
 
+        self.add(RecursiveGroupsInRE('check for recursive groups in OSRC', 'osrc'))
+        self.add(RecursiveGroupsInRE('check for recursive groups in ODST', 'odst'))
+        self.add(RecursiveGroupsInRE('check for recursive groups in OSRV', 'osrv'))
         self.add(EmptyGroupsInRE('check for empty groups in OSRC', 'osrc'))
         self.add(EmptyGroupsInRE('check for empty groups in ODST', 'odst'))
         self.add(EmptyGroupsInRE('check for empty groups in OSRV', 'osrv'))